Bug fix: RMS stats for non-ensemble data series.
This bug was initially pointed out by Dean.
The RMS error of the production series
of the "Truth" (vis-a-vis ref="Noisy" i.e. the obs)
was .23 whereas it should have been sqrt(1e-3) ≈ .03
This was caused by treating the non-ensemble array
as if it was an array whose axis 0 was the ensemble axis.
It also yielded non-0 std.dev. for non-ensemble data series.

This commit also cleans up some of the other related code.